+++ webapp/components/expression/nodes/binary.js
@@ -5,7 +5,7 @@ import * as R from 'ramda'
 import * as Expression from '@core/expressionParser/expression'
 import * as NodeDef from '@core/survey/nodeDef'
 
-import Dropdown from '../../form/dropdown'
+import Dropdown from '../../form/Dropdown'
 import BinaryOperand, { BinaryOperandType } from './binaryOperand'
 import EditButtons from './editButtons'
