+++ webapp/components/expression/nodes/literal.js
@@ -10,7 +10,7 @@ import * as StringUtils from '@core/stringUtils'
 import { SurveyState } from '@webapp/store/survey'
 import * as NodeDefUIProps from '@webapp/loggedin/surveyViews/surveyForm/nodeDefs/nodeDefUIProps'
 import ButtonGroup from '@webapp/components/form/buttonGroup'
-import Dropdown from '../../form/dropdown'
+import Dropdown from '../../form/Dropdown'
 import { useAsyncGetRequest } from '../../hooks'
 import { useI18n, useLang } from '@webapp/store/system'
 import * as ExpressionParser from '../expressionParser'
