+++ webapp/components/survey/NodeDefsSelector/EntitySelector/EntitySelector.js
@@ -7,7 +7,7 @@ import * as Survey from '@core/survey/survey'
 import * as NodeDef from '@core/survey/nodeDef'
 import * as StringUtils from '@core/stringUtils'
 
-import Dropdown from '@webapp/components/form/dropdown'
+import Dropdown from '@webapp/components/form/Dropdown'
 
 const getDropdownItems = (hierarchy, lang) => {
   const entities = []
