+++ webapp/loggedin/surveyViews/category/components/categoryImportSummary.js
@@ -4,7 +4,7 @@ import React from 'react'
 import { connect } from 'react-redux'
 
 import { useI18n } from '@webapp/store/system'
-import Dropdown from '@webapp/components/form/dropdown'
+import Dropdown from '@webapp/components/form/Dropdown'
 import { Modal, ModalBody, ModalFooter } from '@webapp/components/modal'
 
 import * as Category from '@core/survey/category'
