+++ webapp/loggedin/surveyViews/nodeDef/basic/codeProps.js
@@ -4,7 +4,7 @@ import * as R from 'ramda'
 
 import { useI18n } from '@webapp/store/system'
 import { FormItem } from '@webapp/components/form/input'
-import Dropdown from '@webapp/components/form/dropdown'
+import Dropdown from '@webapp/components/form/Dropdown'
 import ButtonGroup from '@webapp/components/form/buttonGroup'
 import CategorySelector from '@webapp/components/survey/CategorySelector'
